:sparkles: (frontend) extended "this" context to handleSubmit
